Implement Parse UserInput
Used regular expressions to parse the user input into command leafs
and connectors. Also if any comments (#) in UserInput everything after
it is ignored.

Still need to check if we want them to be Base* in that function.
Testing for the parsing I did on another file before moving it here.