Current testing and tracing push
This commit is for the purpose of testing the program. Users can pull
from here and see some statements that can be commented our for tracing
seg faults and more.

Tests directory contains scripts that still need working and fixing.
After extensive testing and completing the scripts we can create the
JSON file and the tag and complete our assignment.